Wow, I'm actually doing Work.
Is this really Romy?

+++ Day5_(Einstein's Equation)/ConsoleVersion/src/Einsteins_Equation.java
@@ -17,12 +17,12 @@ public class Einsteins_Equation {
 	}
 	
 	public static void CalculateEnergy() {
-		int massObject;
+		Double massObject;
 		try {
 			System.out.print("Please type in a number.....");
 			String usrInput = scan.nextLine();
 			if (usrInput != null) {
-				massObject = Integer.parseInt(usrInput);
+				massObject = Double.parseDouble(usrInput);
 			}
 			else {
 				return;
